* adjust for show import.

+++ module/caselib/model.php
@@ -563,8 +563,11 @@ class caselibModel extends model
             }
         }
 
-        unlink($this->session->importFile);
-        unset($_SESSION['importFile']);
+        if($this->post->isEndPage)
+        {
+            unlink($this->session->importFile);
+            unset($_SESSION['importFile']);
+        }
     }
 
     /**
